Split IntegrationTests into focused test files
Break the 1229-line monolith into 3 files by concern:
- InlineStylingTests.swift — active + inactive inline styling (370 lines)
- BlockStylingTests.swift — active + inactive block styling, transitions (357 lines)
- EditorFeatureTests.swift — undo/redo, tab indent, font, appearance, full doc (434 lines)
